cleanup pre-post registration
